Added a basic navigator for the mental health playback.
