+++ src/components/Cards/school-networks/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { StyleProp, ViewStyle } from 'react-native';
 
 import { ISubscribedSchoolStats, ISubscribedSchoolGroupStats } from '@covid/core/schools/Schools.dto';
 
@@ -7,9 +8,10 @@ import { SContainerView } from './styles';
 
 type Props = {
   schoolGroups: ISubscribedSchoolGroupStats[];
+  style?: StyleProp<ViewStyle>;
 };
 
-const TransformResponseToUIData = (data: ISubscribedSchoolGroupStats[]): ISubscribedSchoolStats[] => {
+const transformResponseToUIData = (data: ISubscribedSchoolGroupStats[]): ISubscribedSchoolStats[] => {
   return data.reduce((initial: ISubscribedSchoolStats[], group): ISubscribedSchoolStats[] => {
     const school = initial.find((school) => school.id === group.school.id);
 
@@ -36,12 +38,10 @@ const TransformResponseToUIData = (data: ISubscribedSchoolGroupStats[]): ISubscr
 };
 
 function SchoolNetworks(props: Props) {
-  const data: ISubscribedSchoolStats[] = TransformResponseToUIData(props.schoolGroups);
+  const data: ISubscribedSchoolStats[] = transformResponseToUIData(props.schoolGroups);
   return (
-    <SContainerView>
-      {data.map((school, index) => {
-        return school.higher_education ? null : <Card school={school} key={school.id} />;
-      })}
+    <SContainerView style={props.style}>
+      {data.map((school) => (school.higher_education ? null : <Card school={school} key={school.id} />))}
     </SContainerView>
   );
 }
