+++ src/features/index.ts
@@ -1,19 +1,20 @@
+export * from './anniversary';
 export * from './assessment';
-export { CountrySelectScreen } from './CountrySelectScreen';
 export * from './covid-tests';
 export * from './dashboard';
-export { EstimatedCasesScreen } from './EstimatedCasesScreen';
-export { default as LoginScreen } from './login/LoginScreen';
+export * from './diet-study-playback';
 export * from './mental-health';
 export * from './multi-profile';
 export * from './password-reset';
 export * from './patient';
 export * from './register';
 export * from './school-network';
-export { ScreenParamList } from './ScreenParamList';
 export * from './splash';
 export * from './thank-you';
-export * from './validation-study';
 export * from './vaccines';
-export * from './diet-study-playback';
-export * from './anniversary';
+export * from './validation-study';
+export { CountrySelectScreen } from './CountrySelectScreen';
+export { EstimatedCasesScreen } from './EstimatedCasesScreen';
+export { ScreenParamList } from './ScreenParamList';
+export { default as LoginScreen } from './login/LoginScreen';
+export { default as MentalHealthPlayback } from './mental-health-playback';
