+++ src/features/mental-health-playback/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import { BasicPage } from '@covid/components';
+
+export default () => <BasicPage active />;
