--- src/routes/index.ts
@@ -1,4 +1,0 @@
-export { default as MainNavigator } from './main-navigator';
-export { default as MetalHealthNavigator } from './mental-health';
-export { default as DietStudyPlaybackNavigator } from './diet-study-playback-navigator';
-export { default as Anniversary } from './anniversary';
